Add documentation for neo4j implemenation and generic graph pkg
