+++ graph/driver/driver.go
@@ -2,8 +2,6 @@ package driver
 
 import (
 	"context"
-	"errors"
-	"fmt"
 
 	codelistModels "github.com/ONSdigital/dp-code-list-api/models"
 	importModels "github.com/ONSdigital/dp-dimension-importer/model"
@@ -12,32 +10,13 @@ import (
 	obsModels "github.com/ONSdigital/dp-observation-importer/models"
 )
 
-var ErrNotFound = errors.New("not found")
-
-// ErrAttemptsExceededLimit is returned when the number of attempts has reaced
-// the maximum permitted
-type ErrAttemptsExceededLimit struct {
-	WrappedErr error
-}
-
-func (e ErrAttemptsExceededLimit) Error() string {
-	return fmt.Sprintf("number of attempts to execute statement exceeded: %s", e.WrappedErr.Error())
-}
-
-// ErrNonRetriable is returned when the wrapped error type is not retriable
-type ErrNonRetriable struct {
-	WrappedErr error
-}
-
-func (e ErrNonRetriable) Error() string {
-	return fmt.Sprintf("received a non retriable error from neo4j: %s", e.WrappedErr.Error())
-}
-
+// Driver is the base interface any driver implementation must satisfy
 type Driver interface {
 	Close(ctx context.Context) error
 	Healthcheck() (string, error)
 }
 
+// Codelist defines functions to retrieve code list and code nodes
 type CodeList interface {
 	GetCodeLists(ctx context.Context, filterBy string) (*codelistModels.CodeListResults, error)
 	GetCodeList(ctx context.Context, codeListID string) (*codelistModels.CodeList, error)
@@ -48,6 +27,7 @@ type CodeList interface {
 	GetCodeDatasets(ctx context.Context, codeListID, edition string, code string) (*codelistModels.Datasets, error)
 }
 
+// Hierarchy defines functions to create and retrieve generic and instance hierarchy nodes
 type Hierarchy interface {
 	CreateInstanceHierarchyConstraints(ctx context.Context, attempt int, instanceID, dimensionName string) error
 	CloneNodes(ctx context.Context, attempt int, instanceID, codeListID, dimensionName string) error
@@ -64,12 +44,16 @@ type Hierarchy interface {
 	GetHierarchyElement(ctx context.Context, instanceID, dimension, code string) (*hierarchyModels.Response, error)
 }
 
-// Observation provides filtered observation data in CSV rows.
+// Observation defines functions to create and retrieve observation nodes
 type Observation interface {
+	// StreamCSVRows returns a reader which the caller is ultimately responsible for closing
+	// This allows for large volumes of data to be read from a stream without signnificant
+	// memory overhead.
 	StreamCSVRows(ctx context.Context, filter *observation.Filter, limit *int) (observation.StreamRowReader, error)
 	InsertObservationBatch(ctx context.Context, attempt int, instanceID string, observations []*obsModels.Observation, dimensionIDs map[string]string) error
 }
 
+// Instance defines functions to create, update and retrieve details about instances
 type Instance interface {
 	CreateInstanceConstraint(ctx context.Context, i *importModels.Instance) error
 	CreateInstance(ctx context.Context, i *importModels.Instance) error
@@ -81,6 +65,7 @@ type Instance interface {
 	SetInstanceIsPublished(ctx context.Context, instanceID string) error
 }
 
+// Dimension defines functions to create dimension nodes
 type Dimension interface {
 	InsertDimension(ctx context.Context, cache map[string]string, i *importModels.Instance, d *importModels.Dimension) (*importModels.Dimension, error)
 }
