+++ graph/graph.go
@@ -9,6 +9,8 @@ import (
 	"github.com/ONSdigital/dp-graph/mock"
 )
 
+// DB contains the widest possible selection of functionality provided by any
+// implementation of this abstraction
 type DB struct {
 	driver.Driver
 
@@ -19,6 +21,8 @@ type DB struct {
 	driver.Dimension
 }
 
+// Subsets allows a clear and concise way of requesting any combination of
+// functionality by groups of node types
 type Subsets struct {
 	CodeList    bool
 	Hierarchy   bool
@@ -27,26 +31,33 @@ type Subsets struct {
 	Dimension   bool
 }
 
+// NewCodeListStore returns a configured DB containing the CodeList functionality
 func NewCodeListStore(ctx context.Context) (*DB, error) {
 	return New(ctx, Subsets{CodeList: true})
 }
 
+// NewHierarchyStore returns a configured DB containing the Hierarchy functionality
 func NewHierarchyStore(ctx context.Context) (*DB, error) {
 	return New(ctx, Subsets{Hierarchy: true})
 }
 
+// NewObservationStore returns a configured DB containing the Observation functionality
 func NewObservationStore(ctx context.Context) (*DB, error) {
 	return New(ctx, Subsets{Observation: true})
 }
 
+// NewInstanceStore returns a configured DB containing the Instance functionality
 func NewInstanceStore(ctx context.Context) (*DB, error) {
 	return New(ctx, Subsets{Instance: true})
 }
 
+// NewDimensionStore returns a configured DB containing the Dimension functionality
 func NewDimensionStore(ctx context.Context) (*DB, error) {
 	return New(ctx, Subsets{Dimension: true})
 }
 
+// New DB returned according to provided subsets and the environment config
+// satisfying the interfaces requested by the choice of subsets
 func New(ctx context.Context, choice Subsets) (*DB, error) {
 	cfg, err := config.Get()
 	if err != nil {
@@ -99,7 +110,7 @@ func New(ctx context.Context, choice Subsets) (*DB, error) {
 	}, nil
 }
 
-//Test sets flags for managing responses from the Mock driver
+// Test sets flags for managing responses from the Mock driver
 func Test(backend, query, content bool) *mock.Mock {
 	return &mock.Mock{
 		IsBackendReachable: backend,
