+++ neo4j/driver/driver.go
@@ -13,22 +13,27 @@ import (
 //go:generate moq -out ../internal/driver.go -pkg internal . Neo4jDriver
 //go:generate moq -out ../internal/bolt.go -pkg internal . Result
 
+// Result of queries to neo4j - needed for mocking and tests
 type Result bolt.Result
 
+// Neo4jDriver is the interface that wraps basic neo4j driver functionality
 type Neo4jDriver interface {
+	Count(query string) (count int64, err error)
+	Exec(query string, params map[string]interface{}) (bolt.Result, error)
 	Read(query string, mapp mapper.ResultMapper, single bool) error
 	ReadWithParams(query string, params map[string]interface{}, mapp mapper.ResultMapper, single bool) error
 	StreamRows(query string) (*BoltRowReader, error)
-	Count(query string) (count int64, err error)
-	Exec(query string, params map[string]interface{}) (bolt.Result, error)
-	Close(ctx context.Context) error
-	Healthcheck() (string, error)
+
+	driver.Driver
 }
 
+// NeoDriver contains a connection pool and allows basic interaction with the database
 type NeoDriver struct {
 	pool bolt.ClosableDriverPool
 }
 
+// New neo4j closeable connection pool configured with the provided address,
+// connection pool size and timeout
 func New(dbAddr string, size, timeout int) (n *NeoDriver, err error) {
 	pool, err := bolt.NewClosableDriverPoolWithTimeout(dbAddr, size, timeout)
 	if err != nil {
@@ -40,14 +45,19 @@ func New(dbAddr string, size, timeout int) (n *NeoDriver, err error) {
 	}, nil
 }
 
+//Close the contained connection pool
 func (n *NeoDriver) Close(ctx context.Context) error {
 	return n.pool.Close()
 }
 
+// ReadWithParams takes a query, a map of parameters and an indicator of whether
+// a single or list response is expected and writes results into the ResultMapper provided
 func (n *NeoDriver) ReadWithParams(query string, params map[string]interface{}, mapp mapper.ResultMapper, single bool) error {
 	return n.read(query, params, mapp, single)
 }
 
+// ReadWithParams takes a query and an indicator of whether a single or list
+// response is expected and writes results into the ResultMapper provided
 func (n *NeoDriver) Read(query string, mapp mapper.ResultMapper, single bool) error {
 	return n.read(query, nil, mapp, single)
 }
@@ -97,6 +107,9 @@ results:
 	return nil
 }
 
+// StreamRows according to provided query into a Reader and return the Reader and any errors.
+// The Reader will contain reference to the database connection and must be closed
+// by the caller.
 func (n *NeoDriver) StreamRows(query string) (*BoltRowReader, error) {
 	conn, err := n.pool.OpenPool()
 	if err != nil {
@@ -111,10 +124,11 @@ func (n *NeoDriver) StreamRows(query string) (*BoltRowReader, error) {
 	}
 
 	// The connection can only be closed once the results have been read, so the caller is responsible for
-	// calling .CLose() which will ultimately release the connection back into the pool
+	// calling .Close() which will ultimately release the connection back into the pool
 	return NewBoltRowReader(rows, conn), nil
 }
 
+// Count nodes returned by the provided query
 func (n *NeoDriver) Count(query string) (count int64, err error) {
 	c, err := n.pool.OpenPool()
 	if err != nil {
@@ -142,6 +156,7 @@ func (n *NeoDriver) Count(query string) (count int64, err error) {
 	return
 }
 
+// Exec executes the provided query with relevant parameters and returns the response directly
 func (n *NeoDriver) Exec(query string, params map[string]interface{}) (bolt.Result, error) {
 	c, err := n.pool.OpenPool()
 	if err != nil {
